+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Development
 ===========
 - (Fill this out as you fix issues and develop your features).
 - Fix `_cls` that is not set properly in Document constructor (regression) #1950
+- Fix bug in _delta method - Update of a ListField depends on an unrelated dynamic field update #1733
 - Remove deprecated `save()` method and used `insert_one()` #1899
 
 =================